requirements.txt
datadog-api-client==2.26.0
datasets==2.20.0
ddtrace==2.9.3
debugpy==1.8.1
fastapi==0.111.0
gunicorn==22.0
pillow==10.4.0
pre-commit==3.7.1
torch==2.3.0
transformers==4.41.1
